# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TalentAI - AI-Powered Recruitment & Onboarding Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Job {
  id: number;
  title: string;
  description: string;
  requiredSkills?: string[];
  /** @nullable */
  experience?: string | null;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  /** @nullable */
  location?: string | null;
  employmentType: string;
  status: string;
  applicantCount?: number;
  createdAt: string;
}

export interface JobInput {
  title: string;
  description: string;
  requiredSkills?: string[];
  experience?: string;
  salaryMin?: number;
  salaryMax?: number;
  location?: string;
  employmentType: string;
  status?: string;
}

export interface JobUpdate {
  title?: string;
  description?: string;
  requiredSkills?: string[];
  experience?: string;
  salaryMin?: number;
  salaryMax?: number;
  location?: string;
  employmentType?: string;
  status?: string;
}

export interface Candidate {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  location?: string | null;
  skills?: string[];
  /** @nullable */
  experience?: number | null;
  /** @nullable */
  education?: string | null;
  /** @nullable */
  university?: string | null;
  /** @nullable */
  resumeScore?: number | null;
  /** @nullable */
  interviewScore?: number | null;
  /** @nullable */
  finalScore?: number | null;
  /** @nullable */
  recommendation?: string | null;
  status: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  resumeText?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  source?: string | null;
  createdAt: string;
}

export interface CandidateInput {
  name: string;
  email: string;
  phone?: string;
  location?: string;
  skills?: string[];
  experience?: number;
  education?: string;
  university?: string;
  resumeText?: string;
  summary?: string;
  source?: string;
  avatarUrl?: string;
}

export interface CandidateUpdate {
  name?: string;
  email?: string;
  phone?: string;
  location?: string;
  skills?: string[];
  experience?: number;
  education?: string;
  university?: string;
  resumeText?: string;
  summary?: string;
  status?: string;
  resumeScore?: number;
  interviewScore?: number;
  finalScore?: number;
  recommendation?: string;
  avatarUrl?: string;
}

export interface ResumeAnalysis {
  candidateId: number;
  summary: string;
  strengths: string[];
  weaknesses: string[];
  missingSkills: string[];
  recommendedRoles: string[];
  hiringRecommendation: string;
}

export type MatchScoreMatchBreakdown = {
  skillMatch?: number;
  experienceMatch?: number;
  educationMatch?: number;
};

export interface MatchScore {
  candidateId: number;
  jobId: number;
  matchScore: number;
  skillOverlap: string[];
  missingSkills: string[];
  matchBreakdown: MatchScoreMatchBreakdown;
}

export interface Application {
  id: number;
  candidateId: number;
  jobId: number;
  status: string;
  /** @nullable */
  coverLetter?: string | null;
  /** @nullable */
  matchScore?: number | null;
  /** @nullable */
  resumeScore?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  candidate?: Candidate;
  job?: Job;
}

export interface ApplicationInput {
  candidateId: number;
  jobId: number;
  coverLetter?: string;
  notes?: string;
}

export interface ApplicationUpdate {
  status?: string;
  notes?: string;
  matchScore?: number;
  resumeScore?: number;
}

export interface RankedApplication {
  id: number;
  candidateId: number;
  jobId: number;
  status: string;
  /** @nullable */
  matchScore?: number | null;
  /** @nullable */
  resumeScore?: number | null;
  finalScore: number;
  rank: number;
  /** @nullable */
  recommendation?: string | null;
  createdAt: string;
  candidate?: Candidate;
}

export interface Interview {
  id: number;
  candidateId: number;
  jobId: number;
  status: string;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  transcript?: string | null;
  /** @nullable */
  technicalScore?: number | null;
  /** @nullable */
  communicationScore?: number | null;
  /** @nullable */
  relevanceScore?: number | null;
  /** @nullable */
  confidenceScore?: number | null;
  /** @nullable */
  problemSolvingScore?: number | null;
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  aiNotes?: string | null;
  createdAt: string;
  candidate?: Candidate;
  job?: Job;
}

export interface InterviewInput {
  candidateId: number;
  jobId: number;
  scheduledAt?: string;
}

export interface InterviewUpdate {
  status?: string;
  scheduledAt?: string;
  transcript?: string;
  technicalScore?: number;
  communicationScore?: number;
  relevanceScore?: number;
  confidenceScore?: number;
  problemSolvingScore?: number;
  overallScore?: number;
  aiNotes?: string;
}

export interface InterviewTranscript {
  transcript: string;
}

export interface InterviewAnalysis {
  technicalScore: number;
  communicationScore: number;
  relevanceScore: number;
  confidenceScore: number;
  problemSolvingScore: number;
  overallScore: number;
  summary: string;
  recommendation: string;
}

export interface AnalyticsSummary {
  totalApplicants: number;
  shortlisted: number;
  interviewed: number;
  hired: number;
  rejected: number;
  hiringSuccessRate: number;
  avgTimeToHire: number;
  activeJobs: number;
}

export interface FunnelStage {
  stage: string;
  count: number;
}

export interface TimeSeriesPoint {
  date: string;
  count: number;
}

export interface SkillCount {
  skill: string;
  count: number;
}

export interface CopilotMessage {
  message: string;
  /** @nullable */
  context?: string | null;
}

export interface CopilotResponse {
  response: string;
  suggestions?: string[];
}

export interface Onboarding {
  id: number;
  candidateId: number;
  jobId: number;
  status: string;
  offerAccepted?: boolean;
  documentsUploaded?: boolean;
  verificationComplete?: boolean;
  trainingAssigned?: boolean;
  /** @nullable */
  joiningDate?: string | null;
  /** @nullable */
  salary?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  candidate?: Candidate;
  job?: Job;
}

export interface OnboardingInput {
  candidateId: number;
  jobId: number;
  joiningDate?: string;
  salary?: number;
  notes?: string;
}

export interface OnboardingUpdate {
  status?: string;
  offerAccepted?: boolean;
  documentsUploaded?: boolean;
  verificationComplete?: boolean;
  trainingAssigned?: boolean;
  joiningDate?: string;
  salary?: number;
  notes?: string;
}

export type ListJobsParams = {
status?: string;
search?: string;
location?: string;
employmentType?: string;
};

export type ListCandidatesParams = {
search?: string;
status?: string;
jobId?: number;
};

export type ListApplicationsParams = {
jobId?: number;
candidateId?: number;
status?: string;
};

export type GetRankedApplicationsParams = {
jobId: number;
};

export type ListInterviewsParams = {
candidateId?: number;
jobId?: number;
status?: string;
};

export type GetHiringFunnelParams = {
jobId?: number;
};

export type GetTopCandidatesParams = {
limit?: number;
jobId?: number;
};

export type ListOnboardingsParams = {
candidateId?: number;
status?: string;
};