# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TalentAI - AI-Powered Recruitment & Onboarding Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List all candidates
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: jobId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a candidate profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Created candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get a candidate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: Updated candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}/resume-analysis:
    get:
      operationId: getCandidateResumeAnalysis
      tags: [candidates]
      summary: Get AI resume analysis for a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI resume analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeAnalysis"
  /candidates/{id}/match-score/{jobId}:
    get:
      operationId: getCandidateMatchScore
      tags: [candidates]
      summary: Get match score for candidate against a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match score result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchScore"

  # ─── Applications ─────────────────────────────────────────────────────────────
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List all applications
      parameters:
        - name: jobId
          in: query
          schema:
            type: integer
        - name: candidateId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Submit an application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get an application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/ranked:
    get:
      operationId: getRankedApplications
      tags: [applications]
      summary: Get ranked/scored list of applications for a job
      parameters:
        - name: jobId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ranked applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RankedApplication"

  # ─── Interviews ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, title, description, status, employmentType, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        experience:
          type: ["string", "null"]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        location:
          type: ["string", "null"]
        employmentType:
          type: string
        status:
          type: string
        applicantCount:
          type: integer
        createdAt:
          type: string
...
    Candidate:
      type: object
      required: [id, name, email, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        experience:
          type: ["integer", "null"]
        education:
          type: ["string", "null"]
        university:
          type: ["string", "null"]
        resumeScore:
          type: ["number", "null"]
        interviewScore:
          type: ["number", "null"]
        finalScore:
          type: ["number", "null"]
        recommendation:
          type: ["string", "null"]
        status:
          type: string
        avatarUrl:
          type: ["string", "null"]
        resumeText:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        createdAt:
          type: string
    CandidateInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        location:
          type: string
        skills:
          type: array
          items:
            type: string
        experience:
          type: integer
        education:
          type: string
        university:
          type: string
        resumeText:
          type: string
        summary:
          type: string
        source:
          type: string
        avatarUrl:
          type: string
    CandidateUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        location:
          type: string
        skills:
          type: array
          items:
            type: string
        experience:
          type: integer
        education:
          type: string
        university:
          type: string
        resumeText:
          type: string
        summary:
          type: string
        status:
          type: string
        resumeScore:
          type: number
        interviewScore:
          type: number
        finalScore:
          type: number
        recommendation:
          type: string
        avatarUrl:
          type: string
    ResumeAnalysis:
      type: object
      required: [candidateId, summary, strengths, weaknesses, missingSkills, recommendedRoles, hiringRecommendation]
      properties:
        candidateId:
          type: integer
        summary:
          type: string
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        missingSkills:
          type: array
          items:
            type: string
        recommendedRoles:
          type: array
          items:
            type: string
        hiringRecommendation:
          type: string
    MatchScore:
      type: object
      required: [candidateId, jobId, matchScore, skillOverlap, missingSkills, matchBreakdown]
      properties:
        candidateId:
          type: integer
        jobId:
          type: integer
        matchScore:
          type: number
        skillOverlap:
          type: array
          items:
            type: string
        missingSkills:
          type: array
          items:
            type: string
        matchBreakdown:
          type: object
          properties:
            skillMatch:
              type: number
            experienceMatch:
              type: number
            educationMatch:
              type: number

    # ─── Application ──────────────────────────────────────────────────────────
    Application:
      type: object
      required: [id, candidateId, jobId, status, createdAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        jobId:
          type: integer
        status:
          type: string
        coverLetter:
          type: ["string", "null"]
        matchScore:
          type: ["number", "null"]
        resumeScore:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        candidate:
          $ref: "#/components/schemas/Candidate"
        job:
          $ref: "#/components/schemas/Job"
    ApplicationInput:
      type: object
      required: [candidateId, jobId]
      properties:
        candidateId:
          type: integer
        jobId:
          type: integer
        coverLetter:
          type: string
        notes:
          type: string
    ApplicationUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
        matchScore:
          type: number
        resumeScore:
          type: number
    RankedApplication:
      type: object
      required: [id, candidateId, jobId, status, finalScore, rank, createdAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        jobId:
          type: integer
        status:
          type: string
        matchScore:
          type: ["number", "null"]
        resumeScore:
          type: ["number", "null"]
        finalScore:
          type: number
        rank:
          type: integer
        recommendation:
          type: ["string", "null"]
        createdAt:
          type: string
        candidate:
          $ref: "#/components/schemas/Candidate"

    # ─── Interview ────────────────────────────────────────────────────────────
...