# lib/api-zod/src/generated/types/candidateUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TalentAI - AI-Powered Recruitment & Onboarding Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CandidateUpdate {
  name?: string;
  email?: string;
  phone?: string;
  location?: string;
  skills?: string[];
  experience?: number;
  education?: string;
  university?: string;
  resumeText?: string;
  summary?: string;
  status?: string;
  resumeScore?: number;
  interviewScore?: number;
  finalScore?: number;
  recommendation?: string;
  avatarUrl?: string;
}